move dump code into helper file
